+++ nekoyume/Assets/_Scripts/Game/Skill/DoubleAttack.cs
@@ -8,7 +8,7 @@ namespace Nekoyume.Game.Skill
     [Serializable]
     public class DoubleAttack : AttackBase
     {
-        public DoubleAttack(float chance, SkillEffect effect,
+        public DoubleAttack(decimal chance, SkillEffect effect,
             Data.Table.Elemental.ElementalType elemental, int power) : base(chance, effect, elemental, power)
         {
         }
